# docs/openAPI/openapi/payments.yaml
openapi: 3.0.1
info:
  title: EvenToNight - Ticketing API
  version: 1.0.0
  description: API for managing event ticket types, pricing, and availability
servers:
  - url: http://localhost:9050
    description: Local server
  - url: https://ticketing.eventonight.site
    description: Production server
paths:
...
  /internal/events/{eventId}:
    post:
      summary: Create an event (Internal)
      description: |
        Creates an event in the ticketing service.

        **🔒 INTERNAL ENDPOINT**: This endpoint is only accessible to internal microservices.
        It is blocked from external access by Traefik.
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: string
          example: evt_550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventDto'
            example:
              creatorId: "usr_123e4567-e89b-12d3-a456-426614174000"
              date: "2024-12-31T20:00:00Z"
              title: "New Year's Eve Party"
              status: "PUBLISHED"
      responses:
        '201':
          description: Event created successfully
        '400':
          description: Bad Request - Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 400
                message: status must be one of DRAFT, PUBLISHED, COMPLETED, CANCELLED
                error: Bad Request
        '500':
          description: Internal Server Error - Failed to create or update event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 500
                message: Failed to create or update event
                error: Internal Server Error
      tags:
        - Event Internal
...
  /events/{eventId}:
    get:
      summary: Get event by ID
      description: |
        Retrieve details of a specific event. Authentication is optional. Access rules:
        - Unauthenticated users: can only access published events
        - Authenticated users: can access their own events (any status) or published events from others
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: string
          example: evt_550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponseDto'
              example:
                id: evt_550e8400-e29b-41d4-a716-446655440000
                creatorId: usr_123e4567-e89b-12d3-a456-426614174000
                status: PUBLISHED
                date: "2024-12-31T20:00:00Z"
                title: "New Year's Eve Party"
        '403':
          description: Forbidden - Unauthenticated user cannot access a non-published event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: Current user cannot access this event
                error: Forbidden
        '404':
          description: Not Found - Event does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Event with id evt_550e8400-e29b-41d4-a716-446655440000 not found
                error: Not Found
      tags:
        - Event
  /events/{eventId}/ticket-types:
    get:
      summary: Get all ticket types for an event
      description: |
        Retrieve all available ticket types for a specific event.
        Authentication is optional. Access rules:
        - Unauthenticated users: can only access ticket types for published events
        - Authenticated users: can access their own events (any status) or published events from others

        To test with authentication, use the "Authorize" button and provide a Bearer token.
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event
          schema:
            type: string
          example: evt_550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: List of ticket types for the event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventTicketTypeResponseDto'
              example:
                - id: ett_123e4567-e89b-12d3-a456-426614174000
                  eventId: evt_550e8400-e29b-41d4-a716-446655440000
                  type: STANDARD
                  description: "General admission ticket"
                  price: 29.99
                  availableQuantity: 85
                  soldQuantity: 15
                  totalQuantity: 100
                  isSoldOut: false
                - id: ett_456e7890-e89b-12d3-a456-426614174111
                  eventId: evt_550e8400-e29b-41d4-a716-446655440000
                  type: VIP
                  description: "VIP access with backstage pass"
                  price: 99.99
                  availableQuantity: 0
                  soldQuantity: 20
                  totalQuantity: 20
                  isSoldOut: true
        '403':
          description: Forbidden - Unauthenticated user cannot access ticket types for non-published event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: Current user cannot access this event ticket types
                error: Forbidden
        '404':
          description: Not Found - Event does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Event with id evt_550e8400-e29b-41d4-a716-446655440000 not found
                error: Not Found
      tags:
        - Event

    post:
      summary: Create a new ticket type for an event
      description: Define a new ticket type with pricing and quantity for a specific event (creator only)
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          description: ID of the event to create ticket type for
          schema:
            type: string
          example: evt_550e8400-e29b-41d4-a716-446655440000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventTicketTypeDto'
            example:
              type: STANDARD
              description: "General admission ticket"
              price: 29.99
              quantity: 100
      responses:
        '201':
          description: Ticket type created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTicketTypeResponseDto'
              example:
                id: ett_123e4567-e89b-12d3-a456-426614174000
                eventId: evt_550e8400-e29b-41d4-a716-446655440000
                type: STANDARD
                description: "General admission ticket"
                price: 29.99
                availableQuantity: 100
                soldQuantity: 0
                totalQuantity: 100
                isSoldOut: false
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidType:
                  summary: Invalid ticket type
                  value:
                    statusCode: 400
                    message: ["type must be one of: STANDARD, VIP"]
                    error: Bad Request
                invalidPrice:
                  summary: Invalid price
                  value:
                    statusCode: 400
                    message: ["price must not be less than 0"]
                    error: Bad Request
                invalidQuantity:
                  summary: Invalid quantity
                  value:
                    statusCode: 400
                    message: ["quantity must not be less than 1"]
                    error: Bad Request
        '403':
          description: Forbidden - User is not the event creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User ID in token does not match event creator ID
                error: Forbidden
        '409':
          description: Conflict - Ticket type already exists for this event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 409
                message: Ticket type STANDARD already exists for event evt_550e8400-e29b-41d4-a716-446655440000
                error: Conflict
      tags:
        - Event

    delete:
      summary: Delete all ticket types for an event
      description: Delete all ticket types for a specific event (creator only)
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
          example: evt_550e8400-e29b-41d4-a716-446655440000
      responses:
        '204':
          description: Ticket types deleted successfully
        '403':
          description: Forbidden - User is not the event creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User ID in token does not match event creator ID
                error: Forbidden
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Event not found
                error: Not Found
      tags:
        - Event
...
  /ticket-types/{ticketTypeId}:
    get:
      summary: Get a specific ticket type by ID
      description: |
        Retrieve detailed information about a specific ticket type.
        Authentication is optional. Access rules:
        - Unauthenticated users: can only access ticket types for published events
        - Authenticated users: can access their own events (any status) or published events from others

        To test with authentication, use the "Authorize" button and provide a Bearer token.
      security:
        - BearerAuth: []
        - {}
      parameters:
        - name: ticketTypeId
          in: path
          required: true
          description: ID of the ticket type
          schema:
            type: string
          example: ett_123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Ticket type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTicketTypeResponseDto'
              example:
                id: ett_123e4567-e89b-12d3-a456-426614174000
                eventId: evt_550e8400-e29b-41d4-a716-446655440000
                type: STANDARD
                description: "General admission ticket"
                price: 29.99
                availableQuantity: 85
                soldQuantity: 15
                totalQuantity: 100
                isSoldOut: false
        '403':
          description: Forbidden - Current user cannot access ticket type for non-published event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: Current user cannot access this event ticket types
                error: Forbidden
        '404':
          description: Not Found - Ticket type or event does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: EventTicketType with id ett_123e4567-e89b-12d3-a456-426614174000 not found
                error: Not Found
      tags:
        - Event Ticket Types

    put:
      summary: Update ticket type
      description: Update details of a ticket type (creator only)
      security:
        - BearerAuth: []
      parameters:
        - name: ticketTypeId
          in: path
          required: true
          schema:
            type: string
          example: ett_123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventTicketTypeDto'
            example:
              description: "Updated ticket description"
              price: 49.99
              quantity: 150
      responses:
        '200':
          description: Ticket type updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventTicketTypeResponseDto'
              example:
                id: ett_123e4567-e89b-12d3-a456-426614174000
                eventId: evt_550e8400-e29b-41d4-a716-446655440000
                type: STANDARD
                description: "Updated ticket description"
                price: 49.99
                availableQuantity: 150
                soldQuantity: 0
                totalQuantity: 150
                isSoldOut: false
        '403':
          description: Forbidden - User is not the event creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User ID in token does not match event creator ID
                error: Forbidden
        '404':
          description: Ticket type or event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: EventTicketType with id ett_123e4567-e89b-12d3-a456-426614174000 not found
                error: Not Found
      tags:
        - Event Ticket Types

    delete:
      summary: Delete ticket type
      description: Delete a ticket type (creator only)
      security:
        - BearerAuth: []
      parameters:
        - name: ticketTypeId
          in: path
          required: true
          schema:
            type: string
          example: ett_123e4567-e89b-12d3-a456-426614174000
      responses:
        '204':
          description: Ticket type deleted successfully
        '403':
          description: Forbidden - User is not the event creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: User ID in token does not match event creator ID
                error: Forbidden
        '404':
          description: Ticket type or event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: EventTicketType with id ett_123e4567-e89b-12d3-a456-426614174000 not found
                error: Not Found
      tags:
        - Event Ticket Types
  /tickets/{ticketId}:
    get:
      summary: Get ticket details
      description: Retrieve details for a specific ticket (user must own the ticket)
      security:
        - BearerAuth: []
      parameters:
        - name: ticketId
          in: path
          required: true
          description: ID of the ticket
          schema:
            type: string
          example: tkt_123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponseDto'
              example:
                id: tkt_123e4567-e89b-12d3-a456-426614174000
                eventId: evt_550e8400-e29b-41d4-a716-446655440000
                userId: usr_789e0123-e89b-12d3-a456-426614174222
                attendeeName: John Doe
                ticketTypeId: ett_123e4567-e89b-12d3-a456-426614174000
                price: 29.99
                purchaseDate: "2024-03-15T10:30:00Z"
                status: ACTIVE
        '403':
          description: Forbidden - User not authorized to view this ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 403
                message: Not authorized to view this ticket
                error: Forbidden
        '404':
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 404
                message: Ticket not found
                error: Not Found
      tags:
        - Tickets
...
components:
  schemas:
...
    UpdateEventTicketTypeDto:
      type: object
      properties:
        description:
          type: string
          description: Optional description of the ticket type
          example: "Updated ticket description"
        price:
          type: number
          format: float
          description: Updated price (USD)of the ticket
        quantity:
          type: integer
          minimum: 1
          description: Updated total quantity of tickets
          example: 150
      required:
        - price
        - quantity
      example:
        description: "Updated ticket description"
        price: 49.99
        quantity: 150
    EventResponseDto:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the event
          example: evt_550e8400-e29b-41d4-a716-446655440000
        creatorId:
          type: string
          description: ID of the user who created the event
          example: usr_123e4567-e89b-12d3-a456-426614174000
        status:
          type: string
          enum: [DRAFT, PUBLISHED, CANCELLED]
          description: Current status of the event
          example: PUBLISHED
        date:
          type: string
          format: date-time
          nullable: true
          description: Date and time of the event
          example: "2024-12-31T20:00:00Z"
        title:
          type: string
          nullable: true
          description: Title of the event
          example: "New Year's Eve Party"
      required:
        - id
        - creatorId
        - status
    CreateEventDto:
      type: object
      description: Data required to create or update an event
      properties:
        creatorId:
          type: string
          description: ID of the user creating the event
          example: "usr_123e4567-e89b-12d3-a456-426614174000"
        date:
          type: string
          format: date-time
          description: Date and time of the event. Required when status is not DRAFT, optional when status is DRAFT
          example: "2024-12-31T20:00:00Z"
        title:
          type: string
          description: Title of the event (optional, but if provided must not be empty)
          example: "New Year's Eve Party"
        status:
          type: string
          description: Status of the event
          enum: [DRAFT, PUBLISHED, COMPLETED, CANCELLED]
          example: PUBLISHED
      required:
        - creatorId
        - status
      example:
        creatorId: "usr_123e4567-e89b-12d3-a456-426614174000"
        date: "2024-12-31T20:00:00Z"
        title: "New Year's Eve Party"
        status: PUBLISHED
    CreateEventTicketTypeDto:
      type: object
      description: Data required to create a new event ticket type
      properties:
        type:
          type: string
          description: Type of ticket (STANDARD, VIP)
          enum: [STANDARD, VIP]
          example: STANDARD
        description:
          type: string
          description: Description of the ticket type (optional)
          example: "General admission ticket"
        price:
          type: number
          format: float
          description: Price (USD) of the ticket
        quantity:
          type: integer
          minimum: 1
          description: Total quantity of tickets available
          example: 100
      required:
        - type
        - price
        - quantity
      example:
        type: STANDARD
        description: "General admission ticket"
        price: 29.99
        quantity: 100
    EventTicketTypeResponseDto:
      type: object
      description: Complete ticket type information
      properties:
        id:
          type: string
          description: Unique identifier for the ticket type (prefixed with 'ett_')
          example: ett_123e4567-e89b-12d3-a456-426614174000
        eventId:
          type: string
          description: ID of the associated event
          example: evt_550e8400-e29b-41d4-a716-446655440000
        type:
          type: string
          enum: [STANDARD, VIP]
          description: Type of ticket
          example: STANDARD
        description:
          type: string
          description: Description of the ticket type
          example: "General admission ticket"
        price:
          type: number
          format: float
          description: Price (USD) of the ticket
          example: 29.99
        availableQuantity:
          type: integer
          description: Number of tickets still available for purchase
          example: 85
        soldQuantity:
          type: integer
          description: Number of tickets already sold
          example: 15
        totalQuantity:
          type: integer
          description: Total quantity of tickets (available + sold)
          example: 100
        isSoldOut:
          type: boolean
          description: Whether all tickets are sold out
          example: false
      required:
        - id
        - eventId
        - type
        - description
        - price
        - availableQuantity
        - soldQuantity
        - totalQuantity
        - isSoldOut
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        message:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Error message or array of validation errors
          example: "Validation failed"
        error:
          type: string
          description: Error type
          example: Bad Request
      required:
        - statusCode
        - message
...
    TicketResponseDto:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        attendeeName:
          type: string
        ticketTypeId:
          type: string
        price:
          type: number
          format: float
        purchaseDate:
          type: string
          format: date-time
        status:
          type: string
      required:
        - id
        - eventId
        - userId
        - attendeeName
        - ticketTypeId
        - price
        - purchaseDate
        - status
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Enter your JWT token in the format: Bearer <token>'